use numeric project id, add parent relatioship
